float nr double, voor meer precisie

+++ logic/camera.cpp
@@ -7,7 +7,7 @@
 
 camera::camera(unsigned int w, unsigned int h) : height(h), width(w) {}
 
-std::pair<int, int> camera::worldToPixel(const float &x, const float &y) const {
+std::pair<int, int> camera::worldToPixel(const double &x, const double &y) const {
     int pixelX = ((x + 1.f)/ 2.f) * width;
     int pixelY = ((-y + 1.f)/ 2.f) * height;
     return {pixelX, pixelY};
@@ -18,10 +18,10 @@ void camera::setDimensions(unsigned int w, unsigned int h) {
     height = h;
 }
 
-int camera::distanceToPixelsHeight(const float& distance) const {
+int camera::distanceToPixelsHeight(const double& distance) const {
     return std::round((distance)/2 * height) + 1; // +1 om de gap te dichten tussen de entities
 }
 
-int camera::distanceToPixelsWidth(const float& distance) const {
+int camera::distanceToPixelsWidth(const double& distance) const {
     return std::round((distance)/2 * width) + 1; // +1 om de gap te dichten tussen de entities
 }
